+++ cmflib/cmf.py
@@ -56,22 +56,29 @@ from cmflib.metadata_helper import (
 )
 from cmflib.utils.cmf_config import CmfConfig
 from cmflib.utils.helper_functions import get_python_env, change_dir, get_md5_hash
+from cmflib.cmf_server_methods import (
+    merge_created_context, 
+    merge_created_execution, 
+    log_python_env_on_server, 
+    log_dataset_with_version,
+    log_model_with_version, 
+    log_execution_metrics_from_client, 
+    commit_existing_metrics,
+    commit_existing,
+    commit_existing
+)
+
 from cmflib.cmf_commands_wrapper import (
-    _metadata_push,
-    _metadata_pull,
-    _metadata_export,
-    _artifact_pull,
-    _artifact_push,
-    _artifact_pull_single,
-    _cmf_cmd_init,
-    _init_local,
-    _init_minioS3,
-    _init_amazonS3,
-    _init_sshremote,
-    _init_osdfremote,
-    _artifact_list,
-    _pipeline_list,
-    _execution_list,
+    metadata_push,
+    metadata_pull,
+    cmf_init_show,
+    metadata_export,
+    artifact_pull,
+    artifact_push,
+    artifact_pull_single,
+    artifact_list,
+    pipeline_list,
+    execution_list,
 )
 
 class Cmf:
@@ -261,47 +268,6 @@ class Cmf:
         """
         custom_props = {} if custom_properties is None else custom_properties
         pipeline_stage = self.parent_context.name + "/" + pipeline_stage
-        ctx = get_or_create_run_context(
-            self.store, pipeline_stage, custom_props)
-        self.child_context = ctx
-        associate_child_to_parent_context(
-            store=self.store, parent_context=self.parent_context, child_context=ctx
-        )
-        if self.graph:
-            self.driver.create_stage_node(
-                pipeline_stage, self.parent_context, ctx.id, custom_props
-            )
-        return ctx
-
-    def merge_created_context(
-        self, pipeline_stage: str, custom_properties: t.Optional[t.Dict] = None
-    ) -> mlpb.Context:
-        """Merge created context.
-        Every call creates a unique pipeline stage.
-        Created for metadata push purpose.
-        Example:
-
-            ```python
-            #Create context
-            # Import CMF
-            from cmflib.cmf import Cmf
-            from ml_metadata.proto import metadata_store_pb2 as mlpb
-            # Create CMF logger
-            cmf = Cmf(filepath="mlmd", pipeline_name="test_pipeline")
-            # Create context
-            context: mlmd.proto.Context = cmf.merge_created_context(
-                pipeline_stage="Test-env/prepare",
-                custom_properties ={"user-metadata1": "metadata_value"}
-            ```
-            Args:
-                Pipeline_stage: Pipeline_Name/Stage_name.
-                custom_properties: Developers can provide key value pairs with additional properties of the execution that
-                    need to be stored.
-            Returns:
-                Context object from ML Metadata library associated with the new context for this stage.
-        """
-
-        custom_props = {} if custom_properties is None else custom_properties
         ctx = get_or_create_run_context(
             self.store, pipeline_stage, custom_props)
         self.child_context = ctx
@@ -564,127 +530,6 @@ class Cmf:
             )
         return self.execution
 
-    def merge_created_execution(
-        self,
-        execution_type: str,
-        execution_cmd: str,
-        properties: t.Optional[t.Dict] = None,
-        custom_properties: t.Optional[t.Dict] = None,
-        orig_execution_name:str = "",
-        create_new_execution:bool = True
-    ) -> mlpb.Execution:
-        """Merge Created execution.
-        Every call creates a unique execution. Execution can only be created within a context, so
-        [create_context][cmflib.cmf.Cmf.create_context] must be called first.
-        Every call occurs when metadata push or pull is processed. Data from pre-existing executions is used
-        to create new executions with additional data(Required on cmf-server).
-        Example:
-            ```python
-            # Import CMF
-            from cmflib.cmf import Cmf
-            from ml_metadata.proto import metadata_store_pb2 as mlpb
-            # Create CMF logger
-            cmf = Cmf(filepath="mlmd", pipeline_name="test_pipeline")
-            # Create or reuse context for this stage
-            context: mlmd.proto.Context = cmf.merge_created_context(
-                pipeline_stage="prepare",
-                custom_properties ={"user-metadata1": "metadata_value"}
-            )
-            # Create a new execution for this stage run
-            execution: mlmd.proto.Execution = cmf.merge_created_execution(
-                execution_type="Prepare",
-                properties={"Context_Type":""},
-                custom_properties = {"split": split, "seed": seed},
-                orig_execution_name=execution_name
-            )
-            ```
-        Args:
-            execution_type: Type of the execution.(when create_new_execution is False, this is the name of execution)
-            properties: Properties of Execution.
-            custom_properties: Developers can provide key value pairs with additional properties of the execution that
-                need to be stored.
-
-            cmd: command used to run this execution.
-
-            create_new_execution:bool = True, This can be used by advanced users to re-use executions
-                This is applicable, when working with framework code like mmdet, pytorch lightning etc, where the
-                custom call-backs are used to log metrics.
-                if create_new_execution is True(Default), execution_type parameter will be used as the name of the execution type.
-                if create_new_execution is False, if existing execution exist with the same name as execution_type.
-                it will be reused.
-                Only executions created with  create_new_execution as False will have "name" as a property.
-
-
-        Returns:
-            Execution object from ML Metadata library associated with the execution for this stage.
-        """
-        # Initializing the execution related fields
-        properties = {} if properties is None else properties
-        self.metrics = {}
-        self.input_artifacts = []
-        self.execution_label_props = {}
-        custom_props = {} if custom_properties is None else custom_properties
-        # print(custom_props)
-        git_repo = properties.get("Git_Repo", "")
-        git_start_commit = properties.get("Git_Start_Commit", "")
-        #name = properties.get("Name", "")
-        create_new_execution = True
-        execution_name = execution_type
-        #exe.name property is passed as the orig_execution_name.
-        #if name is not an empty string then we are re-using executions
-        if orig_execution_name != "":
-            create_new_execution = False
-            execution_name = orig_execution_name
-
-        self.execution = create_new_execution_in_existing_run_context(
-            store=self.store,
-            execution_type_name=execution_type, # Type field when re-using executions
-            execution_name=execution_name, #Name field if we are re-using executionsname
-                                           #Type field , if creating new executions always
-            context_id=self.child_context.id,
-            execution=execution_cmd,
-            pipeline_id=self.parent_context.id,
-            pipeline_type=self.parent_context.name,
-            git_repo=git_repo,
-            git_start_commit=git_start_commit,
-            custom_properties=custom_props,
-            create_new_execution=create_new_execution
-        )
-
-        uuids = ""
-
-        uuids = self.execution.properties["Execution_uuid"].string_value
-        if uuids:
-            self.execution.properties["Execution_uuid"].string_value = uuids +\
-                ","+properties["Execution_uuid"]
-        else:
-            self.execution.properties["Execution_uuid"].string_value =\
-                properties["Execution_uuid"]
-
-        
-        self.store.put_executions([self.execution])
-        self.execution_name = str(self.execution.id) + "," + execution_type
-        self.execution_command = execution_cmd
-        for k, v in custom_props.items():
-            k = re.sub("-", "_", k)
-            self.execution_label_props[k] = v
-        self.execution_label_props["Execution_Name"] = (
-            execution_type + ":" + str(self.execution.id)
-        )
-        self.execution_label_props["execution_command"] = execution_cmd
-        if self.graph:
-            self.driver.create_execution_node(
-                self.execution_name,
-                self.child_context.id,
-                self.parent_context,
-                execution_cmd,
-                self.execution.id,
-                custom_props,
-            )
-
-        # link the artifact to execution if it exists and creates artifact if it doesn't
-        return self.execution
-
     def log_dvc_lock(self, file_path: str):
         """Used to update the dvc lock file created with dvc run command."""
         print("Entered dvc lock file commit")
@@ -753,87 +598,6 @@ class Cmf:
             self.execution_label_props["git_repo"] = git_repo
             self.execution_label_props["Commit"] = commit
 
-            if self.graph:
-                self.driver.create_env_node(
-                    name,
-                    url,
-                    uri,
-                    "input",
-                    self.execution.id,
-                    self.parent_context,
-                    custom_props,
-                )
-                self.input_artifacts.append(
-                    {
-                        "Name": name,
-                        "Path": url,
-                        "URI": uri,
-                        "Event": "input",
-                        "Execution_Name": self.execution_name,
-                        "Type": "Environment",
-                        "Execution_Command": self.execution_command,
-                        "Pipeline_Id": self.parent_context.id,
-                        "Pipeline_Name": self.parent_context.name,
-                    }
-                )
-                self.driver.create_execution_links(uri, name, "Environment")
-            return artifact
-
-    def log_python_env_on_server(
-            self,
-            url: str,
-            uri: str,
-            props: t.Optional[t.Dict] = None,
-        ) -> mlpb.Artifact:
-            "Used to log the python packages involved in the current execution"
-
-            git_repo = props.get("git_repo", "")
-            name = url
-            existing_artifact = []
-            c_hash = uri
-            commit = props.get("Commit", "")
-            url = url + ":" + c_hash
-            if c_hash and c_hash.strip:
-                existing_artifact.extend(self.store.get_artifacts_by_uri(c_hash))
-
-            if existing_artifact and len(existing_artifact) != 0:
-                existing_artifact = existing_artifact[0]
-                artifact = link_execution_to_artifact(
-                    store=self.store,
-                    execution_id=self.execution.id,
-                    uri=uri,
-                    input_name=url,
-                    event_type=mlpb.Event.Type.INPUT,
-                )
-            else:
-                uri = c_hash if c_hash and c_hash.strip() else str(uuid.uuid1())
-                artifact = create_new_artifact_event_and_attribution(
-                    store=self.store,
-                    execution_id=self.execution.id,
-                    context_id=self.child_context.id,
-                    uri=uri,
-                    name=url,
-                    type_name="Environment",
-                    event_type=mlpb.Event.Type.INPUT,
-                    properties={
-                        "git_repo": str(git_repo),
-                        # passing c_hash value to commit
-                        "Commit": str(commit),
-                        "url": props.get("url", ""),
-                    },
-                    artifact_type_properties={
-                        "git_repo": mlpb.STRING,
-                        "Commit": mlpb.STRING,
-                        "url": mlpb.STRING,
-                    },
-                    milliseconds_since_epoch=int(time.time() * 1000),
-                )
-            custom_props = {}
-            custom_props["git_repo"] = git_repo
-            custom_props["Commit"] = commit
-            self.execution_label_props["git_repo"] = git_repo
-            self.execution_label_props["Commit"] = commit
-            
             if self.graph:
                 self.driver.create_env_node(
                     name,
@@ -1067,143 +831,6 @@ class Cmf:
             put_artifact(self.store, dup_art)
         return dup_artifact
 
-    def log_dataset_with_version(
-        self,
-        url: str,
-        version: str,
-        event: str,
-        props: t.Optional[t.Dict] = None,
-        custom_properties: t.Optional[t.Dict] = None,
-    ) -> mlpb.Artifact:
-        """Logs a dataset when the version (hash) is known.
-           Example: 
-             ```python 
-             artifact: mlpb.Artifact = cmf.log_dataset_with_version( 
-                 url="path/to/dataset", 
-                 version="abcdef",
-                 event="output",
-                 props={ "git_repo": "https://github.com/example/repo",
-                         "url": "/path/in/repo", },
-                 custom_properties={ "custom_key": "custom_value", }, 
-                 ) 
-             ```
-             Args: 
-                url: Path to the dataset. 
-                version: Hash or version identifier for the dataset. 
-                event: Takes arguments `INPUT` or `OUTPUT`. 
-                props: Optional properties for the dataset (e.g., git_repo, url). 
-                custom_properties: Optional custom properties for the dataset.
-             Returns:
-                Artifact object from the ML Protocol Buffers library associated with the new dataset artifact. 
-        """
-
-        props = {} if props is None else props
-        custom_props = {} if custom_properties is None else custom_properties
-        git_repo = props.get("git_repo", "")
-        name = url
-        event_type = mlpb.Event.Type.OUTPUT
-        existing_artifact = []
-        c_hash = version
-        if event.lower() == "input":
-            event_type = mlpb.Event.Type.INPUT
-
-        # dataset_commit = commit_output(url, self.execution.id)
-
-        dataset_commit = version
-        url = url + ":" + c_hash
-        if c_hash and c_hash.strip:
-            existing_artifact.extend(self.store.get_artifacts_by_uri(c_hash))
-
-        # To Do - What happens when uri is the same but names are different
-        if existing_artifact and len(existing_artifact) != 0:
-            existing_artifact = existing_artifact[0]
-
-            # Quick fix- Updating only the name
-            if custom_properties is not None:
-                self.update_existing_artifact(
-                    existing_artifact, custom_properties)
-            uri = c_hash
-            # update url for existing artifact
-            self.update_dataset_url(existing_artifact, props.get("url", ""))
-            artifact = link_execution_to_artifact(
-                store=self.store,
-                execution_id=self.execution.id,
-                uri=uri,
-                input_name=url,
-                event_type=event_type,
-            )
-        else:
-            # if((existing_artifact and len(existing_artifact )!= 0) and c_hash != ""):
-            #   url = url + ":" + str(self.execution.id)
-            uri = c_hash if c_hash and c_hash.strip() else str(uuid.uuid1())
-            artifact = create_new_artifact_event_and_attribution(
-                store=self.store,
-                execution_id=self.execution.id,
-                context_id=self.child_context.id,
-                uri=uri,
-                name=url,
-                type_name="Dataset",
-                event_type=event_type,
-                properties={
-                    "git_repo": str(git_repo),
-                    "Commit": str(dataset_commit),
-                    "url": props.get("url", " "),
-                },
-                artifact_type_properties={
-                    "git_repo": mlpb.STRING,
-                    "Commit": mlpb.STRING,
-                    "url": mlpb.STRING,
-                },
-                custom_properties=custom_props,
-                milliseconds_since_epoch=int(time.time() * 1000),
-            )
-        custom_props["git_repo"] = git_repo
-        custom_props["Commit"] = dataset_commit
-        self.execution_label_props["git_repo"] = git_repo
-        self.execution_label_props["Commit"] = dataset_commit
-
-        if self.graph:
-            self.driver.create_dataset_node(
-                name,
-                url,
-                uri,
-                event,
-                self.execution.id,
-                self.parent_context,
-                custom_props,
-            )
-            if event.lower() == "input":
-                self.input_artifacts.append(
-                    {
-                        "Name": name,
-                        "Path": url,
-                        "URI": uri,
-                        "Event": event.lower(),
-                        "Execution_Name": self.execution_name,
-                        "Type": "Dataset",
-                        "Execution_Command": self.execution_command,
-                        "Pipeline_Id": self.parent_context.id,
-                        "Pipeline_Name": self.parent_context.name,
-                    }
-                )
-                self.driver.create_execution_links(uri, name, "Dataset")
-            else:
-                child_artifact = {
-                    "Name": name,
-                    "Path": url,
-                    "URI": uri,
-                    "Event": event.lower(),
-                    "Execution_Name": self.execution_name,
-                    "Type": "Dataset",
-                    "Execution_Command": self.execution_command,
-                    "Pipeline_Id": self.parent_context.id,
-                    "Pipeline_Name": self.parent_context.name,
-                }
-                self.driver.create_artifact_relationships(
-                    self.input_artifacts, child_artifact, self.execution_label_props
-                )
-        return artifact
-
     # Add the model to dvc do a git commit and store the commit id in MLMD
     def log_model(
         self,
@@ -1376,229 +1003,6 @@ class Cmf:
                 )
         os.chdir(logging_dir)
         return artifact
-
-    # Add the model to dvc do a git commit and store the commit id in MLMD
-    def log_model_with_version(
-        self,
-        path: str,
-        event: str,
-        props=None,
-        custom_properties: t.Optional[t.Dict] = None,
-    ) -> object:
-        """Logs a model when the version(hash) is known
-         The model is added to dvc and the metadata file (.dvc) gets committed to git.
-        Example:
-            ```python
-            artifact: mlmd.proto.Artifact= cmf.log_model_with_version(
-                path="path/to/model.pkl",
-                event="output",
-                props={
-                        "url": "/home/user/local-storage/bf/629ccd5cd008066b72c04f9a918737",
-                        "model_type": "RandomForestClassifier",
-                        "model_name": "RandomForestClassifier:default",
-                        "Commit": "commit 1146dad8b74cae205db6a3132ea403db1e4032e5",
-                        "model_framework": "SKlearn",
-                       },
-                custom_properties={
-                        "uri": "bf629ccd5cd008066b72c04f9a918737",
-                },
-
-            )
-            ```
-        Args:
-            path: Path to the model file.
-            event: Takes arguments `INPUT` OR `OUTPUT`.
-            props: Model artifact properties.
-            custom_properties: The model properties.
-        Returns:
-            Artifact object from ML Metadata library associated with the new model artifact.
-        """
-
-        if custom_properties is None:
-            custom_properties = {}
-        custom_props = {} if custom_properties is None else custom_properties
-        name = re.split("/", path)[-1]
-        event_type = mlpb.Event.Type.OUTPUT
-        existing_artifact = []
-        if event.lower() == "input":
-            event_type = mlpb.Event.Type.INPUT
-
-        # props["commit"] = "" # To do get from incoming data
-        c_hash = props.get("uri", " ")
-        # If connecting to an existing artifact - The name of the artifact is used as path/steps/key
-        model_uri = path + ":" + c_hash
-        # dvc_url = dvc_get_url(path, False)
-        url = props.get("url", "")
-        # uri = ""
-        if c_hash and c_hash.strip():
-            uri = c_hash.strip()
-            existing_artifact.extend(self.store.get_artifacts_by_uri(uri))
-        else:
-            raise RuntimeError("Model commit failed, Model uri empty")
-
-        if (
-            existing_artifact
-            and len(existing_artifact) != 0
-        ):
-            # update url for existing artifact
-            existing_artifact = self.update_model_url(existing_artifact, url)
-            artifact = link_execution_to_artifact(
-                store=self.store,
-                execution_id=self.execution.id,
-                uri=c_hash,
-                input_name=model_uri,
-                event_type=event_type,
-            )
-            model_uri = artifact.name
-        else:
-            uri = c_hash if c_hash and c_hash.strip() else str(uuid.uuid1())
-            model_uri = model_uri + ":" + str(self.execution.id)
-            artifact = create_new_artifact_event_and_attribution(
-                store=self.store,
-                execution_id=self.execution.id,
-                context_id=self.child_context.id,
-                uri=uri,
-                name=model_uri,
-                type_name="Model",
-                event_type=event_type,
-                properties={
-                    "model_framework": props.get("model_framework", ""),
-                    "model_type": props.get("model_type", ""),
-                    "model_name": props.get("model_name", ""),
-                    "Commit": props.get("Commit", ""),
-                    "url": str(url),
-                },
-                artifact_type_properties={
-                    "model_framework": mlpb.STRING,
-                    "model_type": mlpb.STRING,
-                    "model_name": mlpb.STRING,
-                    "Commit": mlpb.STRING,
-                    "url": mlpb.STRING,
-                },
-                custom_properties=custom_props,
-                milliseconds_since_epoch=int(time.time() * 1000),
-            )
-        custom_properties["Commit"] = props.get("Commit", "")
-        custom_props["url"] = url
-        self.execution_label_props["Commit"] = props.get("Commit", "")
-        if self.graph:
-            self.driver.create_model_node(
-                model_uri,
-                uri,
-                event,
-                self.execution.id,
-                self.parent_context,
-                custom_props,
-            )
-            if event.lower() == "input":
-                self.input_artifacts.append(
-                    {
-                        "Name": model_uri,
-                        "URI": uri,
-                        "Event": event.lower(),
-                        "Execution_Name": self.execution_name,
-                        "Type": "Model",
-                        "Execution_Command": self.execution_command,
-                        "Pipeline_Id": self.parent_context.id,
-                        "Pipeline_Name": self.parent_context.name,
-                    }
-                )
-                self.driver.create_execution_links(uri, model_uri, "Model")
-            else:
-                child_artifact = {
-                    "Name": model_uri,
-                    "URI": uri,
-                    "Event": event.lower(),
-                    "Execution_Name": self.execution_name,
-                    "Type": "Model",
-                    "Execution_Command": self.execution_command,
-                    "Pipeline_Id": self.parent_context.id,
-                    "Pipeline_Name": self.parent_context.name,
-                }
-                self.driver.create_artifact_relationships(
-                    self.input_artifacts, child_artifact, self.execution_label_props
-                )
-
-        return artifact
-
-    def log_execution_metrics_from_client(self, metrics_name: str,
-                                         custom_properties: t.Optional[t.Dict] = None) -> mlpb.Artifact:
-        """ Logs execution metrics from a client.
-            Data from pre-existing metrics from client side is used to create identical metrics on server side. 
-            Example: 
-              ```python 
-              artifact: mlpb.Artifact = cmf.log_execution_metrics_from_client( 
-                      metrics_name="example_metrics:uri:123", 
-                      custom_properties={"custom_key": "custom_value"}, 
-                      )
-              ``` 
-              Args: 
-                 metrics_name: Name of the metrics in the format "name:uri:execution_id". 
-                 custom_properties: Optional custom properties for the metrics. 
-              Returns: 
-                 Artifact object from the ML Protocol Buffers library associated with the metrics artifact.
-        """
-
-        metrics = None
-        custom_props = {} if custom_properties is None else custom_properties
-        existing_artifact = []
-        name_tokens = metrics_name.split(":")
-        if name_tokens and len(name_tokens) > 2:
-            name = name_tokens[0]
-            uri = name_tokens[1]
-            execution_id = name_tokens[2]
-        else:
-            print(f"Error : metrics name {metrics_name} is not in the correct format")
-            return 
-
-        #we need to add the execution id to the metrics name
-        new_metrics_name = f"{name}:{uri}:{str(self.execution.id)}"
-        existing_artifacts = self.store.get_artifacts_by_uri(uri)
-
-        existing_artifact = existing_artifacts[0] if existing_artifacts else None
-        # Didn't understand this, 
-        # and in case of step_metrics should we follow this logic or dataset's logic or does it even matter
-        if not existing_artifact or \
-           ((existing_artifact) and not
-            (existing_artifact.name == new_metrics_name)):  #we need to add the artifact otherwise its already there 
-            metrics = create_new_artifact_event_and_attribution(
-            store=self.store,
-            execution_id=self.execution.id,
-            context_id=self.child_context.id,
-            uri=uri,
-            name=new_metrics_name,
-            type_name="Metrics",
-            event_type=mlpb.Event.Type.OUTPUT,
-            properties={"metrics_name": metrics_name},
-            artifact_type_properties={"metrics_name": mlpb.STRING},
-            custom_properties=custom_props,
-            milliseconds_since_epoch=int(time.time() * 1000),
-        )
-            
-            if self.graph:
-                # To do create execution_links
-                self.driver.create_metrics_node(
-                    metrics_name,
-                    uri,
-                    "output",
-                    self.execution.id,
-                    self.parent_context,
-                    custom_props,
-                )
-                child_artifact = {
-                    "Name": metrics_name,
-                    "URI": uri,
-                    "Event": "output",
-                    "Execution_Name": self.execution_name,
-                    "Type": "Metrics",
-                    "Execution_Command": self.execution_command,
-                    "Pipeline_Id": self.parent_context.id,
-                    "Pipeline_Name": self.parent_context.name,
-                }
-                self.driver.create_artifact_relationships(
-                    self.input_artifacts, child_artifact, self.execution_label_props
-                )
-        return metrics
 
 
     def log_execution_metrics(
@@ -1812,84 +1216,6 @@ class Cmf:
             )
 
         os.chdir(logging_dir)
-        return metrics
-
-    def commit_existing_metrics(self, metrics_name: str, uri: str, props: t.Optional[t.Dict] = None, custom_properties: t.Optional[t.Dict] = None):
-        """
-        Commits existing metrics associated with the given URI to MLMD.
-        Example:
-        ```python
-           artifact: mlpb.Artifact = cmf.commit_existing_metrics("existing_metrics", "abc123",
-           {"custom_key": "custom_value"})
-        ```
-        Args:
-           metrics_name: Name of the metrics.
-           uri: Unique identifier associated with the metrics.
-           custom_properties: Optional custom properties for the metrics.
-        Returns:
-           Artifact object from the ML Protocol Buffers library associated with the existing metrics artifact.
-        """
-
-        custom_props =  {} if custom_properties is None else custom_properties
-        c_hash = uri.strip()
-        existing_artifact = []
-        existing_artifact.extend(self.store.get_artifacts_by_uri(c_hash))
-        if (existing_artifact
-            and len(existing_artifact) != 0 ):
-            metrics = link_execution_to_artifact(
-                store=self.store,
-                execution_id=self.execution.id,
-                uri=c_hash,
-                input_name=metrics_name,
-                event_type=mlpb.Event.Type.OUTPUT,
-            )
-        else:
-            metrics = create_new_artifact_event_and_attribution(
-                store=self.store,
-                execution_id=self.execution.id,
-                context_id=self.child_context.id,
-                uri=uri,
-                name=metrics_name,
-                type_name="Step_Metrics",
-                event_type=mlpb.Event.Type.OUTPUT,
-                properties={
-                    # passing uri value to commit
-                    "Commit": props.get("Commit", ""),
-                    "url": props.get("url", ""),
-                },
-                artifact_type_properties={
-                    "Commit": mlpb.STRING,
-                    "url": mlpb.STRING,
-                },
-                custom_properties=custom_props,
-                milliseconds_since_epoch=int(time.time() * 1000),
-            )
-        
-        metrics_commit = props.get("Commit", "")
-        custom_props["Commit"] = metrics_commit
-        self.execution_label_props["Commit"] = metrics_commit
-
-        if self.graph:
-            self.driver.create_step_metrics_node(
-                metrics_name,
-                uri,
-                "output",
-                self.execution.id,
-                self.parent_context,
-                custom_props,
-            )
-            child_artifact = {
-                "Name": metrics_name,
-                "URI": uri,
-                "Event": "output",
-                "Execution_Name": self.execution_name,
-                "Type": "Step_Metrics",
-                "Execution_Command": self.execution_command,
-                "Pipeline_Id": self.parent_context.id,
-            }
-            self.driver.create_artifact_relationships(
-                self.input_artifacts, child_artifact, self.execution_label_props
-            )
         return metrics
 
 
@@ -2143,53 +1469,6 @@ class Cmf:
             os.chdir(logging_dir)
             return slice
 
-        # commit existing dataslice to server
-        def commit_existing(self, uri: str, props: t.Optional[t.Dict] = None, custom_properties: t.Optional[t.Dict] = None) -> None:
-            custom_props = {} if custom_properties is None else custom_properties
-            c_hash = uri.strip()
-            dataslice_commit = c_hash
-            existing_artifact = []
-            if c_hash and c_hash.strip():
-                existing_artifact.extend(
-                    self.writer.store.get_artifacts_by_uri(c_hash))
-            if existing_artifact and len(existing_artifact) != 0:
-                print("Adding to existing data slice")
-                 # Haven't added event type in this if cond, is it not needed??
-                slice = link_execution_to_input_artifact(
-                    store=self.writer.store,
-                    execution_id=self.writer.execution.id,
-                    uri=c_hash,
-                    input_name=self.name,
-                )
-            else:
-                slice = create_new_artifact_event_and_attribution(
-                    store=self.writer.store,
-                    execution_id=self.writer.execution.id,
-                    context_id=self.writer.child_context.id,
-                    uri=c_hash,
-                    name=self.name,
-                    type_name="Dataslice",
-                    event_type=mlpb.Event.Type.OUTPUT,
-                    properties={
-                        "git_repo": props.get("git_repo", ""),
-                        "Commit": props.get("Commit", ""),
-                        "url": props.get("url", " "),
-                    },
-                    artifact_type_properties={
-                        "git_repo": mlpb.STRING,
-                        "Commit": mlpb.STRING,
-                        "url": mlpb.STRING,
-                    },
-                    custom_properties=custom_props,
-                    milliseconds_since_epoch=int(time.time() * 1000),
-                )
-            custom_props["git_repo"] = props.get("git_repo", "")
-            custom_props["Commit"] = props.get("Commit", "")
-            if self.writer.graph:
-                self.writer.driver.create_dataslice_node(
-                    self.name, self.name, c_hash, self.data_parent, custom_props
-                )
-            return slice
 
 
 #        """Temporary code"""
@@ -2204,371 +1483,27 @@ class Cmf:
 #                print(last)
 #                os.symlink(str(index), slicedir + "/ " + last)
 
+# these are cmf logging api needed for server and defined in cmf_server_methods.py file 
 
-def metadata_push(pipeline_name: str, filepath = "./mlmd", tensorboard_path: str = "", execution_id: str = ""):
-    """ Pushes MLMD file to CMF-server.
-    Example:
-    ```python
-         result = metadata_push("example_pipeline", "mlmd_file", "3")
-    ```
-    Args:
-        pipeline_name: Name of the pipeline.
-        filepath: Path to the MLMD file.
-        execution_id: Optional execution ID.
-        tensorboard_path: Path to tensorboard logs.
+Cmf.merge_created_context = merge_created_context
+Cmf.merge_created_execution = merge_created_execution
+Cmf.log_python_env_on_server = log_python_env_on_server
+Cmf.log_dataset_with_version = log_dataset_with_version
+Cmf.log_model_with_version = log_model_with_version
+Cmf.log_execution_metrics_from_client =  log_execution_metrics_from_client
+Cmf.commit_existing_metrics = commit_existing_metrics
+#log_metrics_from_client
+Cmf.DataSlice.commit_existing = commit_existing
+# log_dataslice_from_client
 
-    Returns:
-        Response output from the _metadata_push function.
-    """
-    # Required arguments:  pipeline_name
-    # Optional arguments: Execution_ID, filepath (mlmd file path, tensorboard_path
-    output = _metadata_push(pipeline_name, filepath, execution_id, tensorboard_path)
-    return output
+Cmf.metadata_push = metadata_push
+Cmf.metadata_pull = metadata_pull
+Cmf.metadata_export = metadata_export
+Cmf.artifact_pull = artifact_pull
+Cmf.artifact_pull_single = artifact_pull_single
+Cmf.artifact_push = artifact_push
+Cmf.cmf_init_show = cmf_init_show
+Cmf.pipeline_list = pipeline_list
+Cmf.execution_list = execution_list
+Cmf.artifact_list = artifact_list
 
-def metadata_pull(pipeline_name: str, filepath = "./mlmd", execution_id: str = ""):
-    """ Pulls MLMD file from CMF-server. 
-     Example: 
-     ```python 
-          result = metadata_pull("example_pipeline", "./mlmd_directory", "execution_123") 
-     ``` 
-     Args: 
-        pipeline_name: Name of the pipeline. 
-        filepath: File path to store the MLMD file. 
-        execution_id: Optional execution ID. 
-     Returns: 
-        Message from the _metadata_pull function. 
-     """
-    # Required arguments:  pipeline_name 
-    #Optional arguments: Execution_ID, filepath(file path to store mlmd file) 
-    output = _metadata_pull(pipeline_name, filepath, execution_id)
-    return output
-
-def metadata_export(pipeline_name: str, jsonfilepath: str = "", filepath = "./mlmd"):
-    """ Export local mlmd's metadata in json format to a json file. 
-     Example: 
-     ```python 
-          result = metadata_pull("example_pipeline", "./jsonfile", "./mlmd_directory") 
-     ``` 
-     Args: 
-        pipeline_name: Name of the pipeline. 
-        jsonfilepath: File path of json file. 
-        filepath: File path to store the MLMD file. 
-     Returns: 
-        Message from the _metadata_pull function. 
-     """
-    # Required arguments:  pipeline_name 
-    #Optional arguments: jsonfilepath, filepath(file path to store mlmd file) 
-    output = _metadata_export(pipeline_name, jsonfilepath, filepath)
-    return output
-
-def artifact_pull(pipeline_name: str, filepath = "./mlmd"):
-    """ Pulls artifacts from the initialized repository.
-
-    Example:
-    ```python
-         result = artifact_pull("example_pipeline", "./mlmd_directory")
-    ```
-
-    Args:
-        pipeline_name: Name of the pipeline.
-        filepath: Path to store artifacts.
-    Returns:
-        Output from the _artifact_pull function.
-    """
-
-    # Required arguments: Pipeline_name
-    # Optional arguments: filepath( path to store artifacts)
-    output = _artifact_pull(pipeline_name, filepath)
-    return output
-
-def artifact_pull_single(pipeline_name: str, filepath: str, artifact_name: str):
-    """ Pulls a single artifact from the initialized repository. 
-    Example: 
-    ```python 
-        result = artifact_pull_single("example_pipeline", "./mlmd_directory", "example_artifact") 
-    ```
-    Args: 
-       pipeline_name: Name of the pipeline. 
-       filepath: Path to store the artifact. 
-       artifact_name: Name of the artifact. 
-    Returns:
-       Output from the _artifact_pull_single function. 
-    """
-
-    # Required arguments: Pipeline_name
-    # Optional arguments: filepath( path to store artifacts), artifact_name
-    output = _artifact_pull_single(pipeline_name, filepath, artifact_name)
-    return output
-
-def artifact_push(pipeline_name: str, filepath = "./mlmd"):
-    """ Pushes artifacts to the initialized repository.
-
-    Example:
-    ```python
-         result = artifact_push("example_pipeline", "./mlmd_directory")
-    ```
-    Args: 
-       pipeline_name: Name of the pipeline. 
-       filepath: Path to store the artifact. 
-    Returns:
-        Output from the _artifact_push function.
-    """
-
-    output = _artifact_push(pipeline_name, filepath)
-    return output
-
-def cmf_init_show():
-    """ Initializes and shows details of the CMF command. 
-    Example: 
-    ```python 
-         result = cmf_init_show() 
-    ``` 
-    Returns: 
-       Output from the _cmf_cmd_init function. 
-    """
-
-    output=_cmf_cmd_init()
-    return output
-
-def cmf_init(type: str = "",
-        path: str = "",
-        git_remote_url: str = "",
-        cmf_server_url: str = "",
-        neo4j_user: str = "",
-        neo4j_password: str = "",
-        neo4j_uri: str = "",
-        url: str = "",
-        endpoint_url: str = "",
-        access_key_id: str = "",
-        secret_key: str = "",
-        session_token: str = "",
-        user: str = "",
-        password: str = "",
-        port: int = 0,
-        osdf_path: str = "",
-        key_id: str = "",
-        key_path: str = "",
-        key_issuer: str = "",
-         ):
-
-    """ Initializes the CMF configuration based on the provided parameters. 
-    Example:
-    ```python
-       cmf_init( type="local", 
-                 path="/path/to/re",
-                 git_remote_url="git@github.com:user/repo.git",
-                 cmf_server_url="http://cmf-server"
-                 neo4j_user", 
-                 neo4j_password="password",
-                 neo4j_uri="bolt://localhost:76"
-               )
-    ```
-    Args: 
-       type: Type of repository ("local", "minioS3", "amazonS3", "sshremote")
-       path: Path for the local repository. 
-       git_remote_url: Git remote URL for version control.
-       cmf_server_url: CMF server URL.
-       neo4j_user: Neo4j database username.
-       neo4j_password: Neo4j database password.
-       neo4j_uri: Neo4j database URI.
-       url: URL for MinioS3 or AmazonS3.
-       endpoint_url: Endpoint URL for MinioS3.
-       access_key_id: Access key ID for MinioS3 or AmazonS3.
-       secret_key: Secret key for MinioS3 or AmazonS3. 
-       session_token: Session token for AmazonS3.
-       user: SSH remote username.
-       password: SSH remote password. 
-       port: SSH remote port
-    Returns:
-       Output based on the initialized repository type.
-    """
-
-    if type == "":
-        return print("Error: Type is not provided")
-    if type not in ["local","minioS3","amazonS3","sshremote","osdfremote"]:
-        return print("Error: Type value is undefined"+ " "+type+".Expected: "+",".join(["local","minioS3","amazonS3","sshremote","osdfremote"]))
-
-    if neo4j_user != "" and  neo4j_password != "" and neo4j_uri != "":
-        pass
-    elif neo4j_user == "" and  neo4j_password == "" and neo4j_uri == "":
-        pass
-    else:
-        return print("Error: Enter all neo4j parameters.") 
-
-    args={'path': path,
-        'git_remote_url': git_remote_url,
-        'url': url,
-        'endpoint_url': endpoint_url,
-        'access_key_id': access_key_id,
-        'secret_key': secret_key,
-        'session_token': session_token,
-        'user': user,
-        'password': password,
-        'osdf_path': osdf_path,
-        'key_id': key_id,
-        'key_path': key_path, 
-        'key-issuer': key_issuer,
-        }
-
-    status_args=non_related_args(type, args)
-
-    if type == "local" and path != "" and  git_remote_url != "" :
-        """Initialize local repository"""
-        output = _init_local(
-            path, git_remote_url, cmf_server_url, neo4j_user, neo4j_password, neo4j_uri
-        )
-        if status_args != []:
-            print("There are non-related arguments: "+",".join(status_args)+".Please remove them.")
-        return output
-         
-    elif type == "minioS3" and url != "" and endpoint_url != "" and access_key_id != "" and secret_key != "" and git_remote_url != "":
-        """Initialize minioS3 repository"""
-        output = _init_minioS3(
-            url,
-            endpoint_url,
-            access_key_id,
-            secret_key,
-            git_remote_url,
-            cmf_server_url,
-            neo4j_user,
-            neo4j_password,
-            neo4j_uri,
-        )
-        if status_args != []:
-            print("There are non-related arguments: "+",".join(status_args)+".Please remove them.")
-        return output
-
-    elif type == "amazonS3" and url != "" and access_key_id != "" and secret_key != "" and git_remote_url != "":
-        """Initialize amazonS3 repository"""
-        output = _init_amazonS3(
-            url,
-            access_key_id,
-            secret_key,
-            session_token,
-            git_remote_url,
-            cmf_server_url,
-            neo4j_user,
-            neo4j_password,
-            neo4j_uri,
-        )
-        if status_args != []:
-            print("There are non-related arguments: "+",".join(status_args)+".Please remove them.")
-
-        return output
-
-    elif type == "sshremote" and path != "" and user != "" and port != 0 and password != "" and git_remote_url != "":
-        """Initialize sshremote repository"""
-        output = _init_sshremote(
-            path,
-            user,
-            port,
-            password,
-            git_remote_url,
-            cmf_server_url,
-            neo4j_user,
-            neo4j_password,
-            neo4j_uri,
-        )
-        if status_args != []:
-            print("There are non-related arguments: "+",".join(status_args)+".Please remove them.")
-
-        return output
-
-    elif type == "osdfremote" and osdf_path != "" and key_id != "" and key_path != 0 and key_issuer != "" and git_remote_url != "":
-        """Initialize osdfremote repository"""
-        output = _init_osdfremote(
-            osdf_path,
-            key_id,
-            key_path,
-            key_issuer,
-            git_remote_url,
-            cmf_server_url,
-            neo4j_user,
-            neo4j_password,
-            neo4j_uri,
-        )
-        if status_args != []:
-            print("There are non-related arguments: "+",".join(status_args)+".Please remove them.")
-
-        return output
-
-    else:
-        print("Error: Enter all arguments")
-
-
-def non_related_args(type : str, args : dict):
-    available_args=[i for i, j in args.items() if j != ""]
-    local=["path", "git_remote_url"]
-    minioS3=["url", "endpoint_url", "access_key_id", "secret_key", "git_remote_url"]
-    amazonS3=["url", "access_key_id", "secret_key", "session_token", "git_remote_url"]
-    sshremote=["path", "user", "port", "password", "git_remote_url"]
-    osdfremote=["osdf_path", "key_id", "key_path", "key-issuer", "git_remote_url"]
-
-
-    dict_repository_args={"local" : local, "minioS3" : minioS3, "amazonS3" : amazonS3, "sshremote" : sshremote}
-    
-    for repo,arg in dict_repository_args.items():
-        if repo ==type:
-            non_related_args=list(set(available_args)-set(dict_repository_args[repo]))
-    return non_related_args
-
-
-def pipeline_list(filepath = "./mlmd"):
-    """ Display list of pipline for current mlmd.
-
-    Example:
-    ```python
-         result = _pipeline_list("./mlmd_directory")
-    ```
-
-    Args:
-        filepath: File path to store the MLMD file. 
-    Returns:
-        Output from the _pipeline_list function.
-    """
-
-    # Optional arguments: filepath( path to store the MLMD file)
-    output = _pipeline_list(filepath)
-    return output
-
-
-def execution_list(pipeline_name: str, filepath = "./mlmd", execution_id: str = "", long = True):
-    """ Display list of execution for given pipeline.
-    Example: 
-    ```python 
-        result = _execution_list("example_pipeline", "./mlmd_directory", "example_execution_id", "long") 
-    ```
-    Args: 
-       pipeline_name: Name of the pipeline. 
-       filepath: Path to store the mlmd file. 
-       execution_id: Executions for particular execution id. 
-       long: Detailed summary regarding execution.
-    Returns:
-       Output from the _execution_list function. 
-    """
-
-    # Required arguments: pipeline_name
-    # Optional arguments: filepath( path to store mlmd file), execution_id, long
-    output = _execution_list(pipeline_name, filepath, execution_id, long)
-    return output
-
-
-def artifact_list(pipeline_name: str, filepath = "./mlmd", artifact_name: str = "", long = True):
-    """ Display list of artifact for given pipeline.
-    Example: 
-    ```python 
-        result = _artifact_list("example_pipeline", "./mlmd_directory", "example_artifact_name", "long") 
-    ```
-    Args: 
-       pipeline_name: Name of the pipeline. 
-       filepath: Path to store the mlmd file. 
-       artifact_name: Artifacts for particular artifact name. 
-       long: Detailed summary regarding artifact.
-    Returns:
-       Output from the _artifact_list function. 
-    """
-
-    # Required arguments: pipeline_name
-    # Optional arguments: filepath( path to store mlmd file), artifact_name, long
-    output = _artifact_list(pipeline_name, filepath, artifact_name, long)
-    return output
